# docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.CreateTeamRequest:
    properties:
      address:
        type: string
      city:
        type: string
      founded_year:
        minimum: 1
        type: integer
      logo_url:
        type: string
      name:
        type: string
    required:
    - address
    - city
    - founded_year
    - name
    type: object
  internal_handler.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
      success:
        type: boolean
    type: object
...
  internal_handler.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  internal_handler.RevertRequest:
    properties:
      target_version:
        type: integer
    required:
    - target_version
    type: object
  internal_handler.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_handler.UpdateTeamRequest:
    properties:
      address:
        type: string
      city:
        type: string
      founded_year:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      version:
        type: integer
    required:
    - version
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing amateur football teams, players, matches, and
    reports for Perusahaan XYZ.
  title: Kickbase Football Team Management API
  version: "1.0"
paths:
...
  /api/teams:
    get:
      description: Get a paginated list of teams with optional search and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by team name
        in: query
        name: name
        type: string
      - description: Filter by headquarters city
        in: query
        name: city
        type: string
      - default: created_at
        description: Sort field (name, city, founded_year, created_at)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.PaginatedResponse'
      summary: List all teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Create a new football team with the provided details
      parameters:
      - description: Team data
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Create a new team
      tags:
      - Teams
  /api/teams/{id}:
    delete:
      description: Soft delete a team (only if no active players)
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Delete a team
      tags:
      - Teams
    get:
      description: Get a single team by its ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get a team by ID
      tags:
      - Teams
    put:
      consumes:
      - application/json
      description: Update an existing team's information
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Team data with version
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Update a team
      tags:
      - Teams
  /api/teams/{id}/history:
    get:
      description: Get the version history of a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
      summary: Get team history
      tags:
      - Teams
  /api/teams/{id}/revert:
    post:
      consumes:
      - application/json
      description: Revert a team to a specific previous version
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Target version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RevertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Revert team to previous version
      tags:
      - Teams
produces:
- application/json
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"